requirements.txt
# Core ML (torch and torchvision are in the base image)
timm==0.6.13
numpy==1.24.3

# Image processing
Pillow==10.0.0
opencv-python-headless==4.8.0.76
albumentations==1.3.0

# Web interface
gradio==4.20.0
# Align FastAPI stack to Pydantic v2 (compatible with Gradio 4.20.0)
fastapi==0.110.0
starlette==0.36.3
pydantic==2.6.4

# Optional (uncomment if needed)
pandas==2.0.3
# scikit-learn==1.3.0
# matplotlib==3.7.1
# seaborn==0.12.2
# tqdm==4.65.0
# wandb==0.15.8

# Training framework and metrics
pytorch-lightning==2.3.2
torchmetrics==1.4.0

# Explainability
git+https://github.com/jacobgil/pytorch-grad-cam.git@master